+++ lib/src/widgets/image_provider/image_provider_mobile.dart
@@ -1,10 +1,13 @@
 import 'dart:io';
+import 'dart:typed_data';
 
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 
-ImageProvider getImageProvider(String url) {
-  if (url.startsWith('http')) {
+ImageProvider getImageProvider(String url, {Uint8List? fileBytes}) {
+  if (fileBytes != null) {
+    return MemoryImage(fileBytes); // bytes (Flutter Web/Desktop)
+  } else if (url.startsWith('http') || url.startsWith('blob')) {
     return CachedNetworkImageProvider(url);
   } else if (url.startsWith('assets')) {
     return AssetImage(url);