+++ lib/src/widgets/image_provider/image_provider_web.dart
@@ -1,12 +1,16 @@
+import 'dart:typed_data';
+
 import 'package:flutter/material.dart';
 
-ImageProvider getImageProvider(String url) {
-  if (url.startsWith('http')) {
+ImageProvider getImageProvider(String url, {Uint8List? fileBytes}) {
+  if (fileBytes != null) {
+    return MemoryImage(fileBytes); // bytes (Flutter Web/Desktop)
+  } else if (url.startsWith('http') || url.startsWith('blob')) {
     return NetworkImage(url);
   } else {
     return const AssetImage(
       'assets/placeholder.png',
-      package: 'dash_chat_2',
+      package: 'dash_chat_custom',
     );
   }
 }