+++ lib/src/widgets/message_row/media_container.dart
@@ -1,22 +1,54 @@
-part of '../../../dash_chat_2.dart';
+part of '../../../dash_chat_custom.dart';
 
 /// @nodoc
 class MediaContainer extends StatelessWidget {
   const MediaContainer({
+    required this.maxWidth,
+    required this.maxHeight,
     required this.message,
-    required this.isOwnMessage,
     this.messageOptions = const MessageOptions(),
+    this.previousMessage,
+    this.nextMessage,
+    this.isOwnMessage = false,
+    this.isPreviousSameAuthor = false,
+    this.isNextSameAuthor = false,
+    this.isAfterDateSeparator = false,
+    this.isBeforeDateSeparator = false,
     Key? key,
   }) : super(key: key);
 
+  /// The Chat media max width (default is full width)
+  final double maxWidth;
+
+  /// The Chat media max height (default is full height)
+  final double maxHeight;
+  
   /// Message that contains the media to show
   final ChatMessage message;
+
+  /// Options to customize the behaviour and design of the messages
+  final MessageOptions messageOptions;
+
+  /// Previous message in the list
+  final ChatMessage? previousMessage;
+
+  /// Next message in the list
+  final ChatMessage? nextMessage;
 
   /// If the message is from the current user
   final bool isOwnMessage;
 
-  /// Options to customize the behaviour and design of the messages
-  final MessageOptions messageOptions;
+  /// If the previous message is from the same author as the current one
+  final bool isPreviousSameAuthor;
+
+  /// If the next message is from the same author as the current one
+  final bool isNextSameAuthor;
+
+  /// If the message is preceded by a date separator
+  final bool isAfterDateSeparator;
+
+  /// If the message is before by a date separator
+  final bool isBeforeDateSeparator;
 
   /// Get the right media widget according to its type
   Widget _getMedia(
@@ -44,26 +76,47 @@ class MediaContainer extends StatelessWidget {
         return Stack(
           alignment: AlignmentDirectional.bottomEnd,
           children: <Widget>[
-            Image(
-              height: height,
-              width: width,
-              fit: BoxFit.cover,
-              alignment: isOwnMessage ? Alignment.topRight : Alignment.topLeft,
-              image: getImageProvider(media.url),
-            ),
+            if (media.isUploading)
+              Image(
+                height: height,
+                width: width,
+                fit: BoxFit.cover,
+                alignment: isOwnMessage ? Alignment.topRight : Alignment.topLeft,
+                image: getImageProvider(media.url, fileBytes: media.fileBytes),
+              ),
+
+            if (!media.isUploading)
+              WidgetZoomPro(
+                hoverCursor: SystemMouseCursors.click,
+                enableEmbeddedView: false,
+                heroAnimationTag: const Uuid().v1(),
+                closeFullScreenImageOnEscape: true,
+                zoomWidget: Image(
+                  height: height,
+                  width: width,
+                  fit: BoxFit.cover,
+                  alignment: isOwnMessage ? Alignment.topRight : Alignment.topLeft,
+                  image: getImageProvider(media.url, fileBytes: media.fileBytes),
+                ),
+              ),
+
             if (media.isUploading) loading
           ],
         );
       default:
+        //file
         return TextContainer(
+          maxWidth: maxWidth,
+          maxHeight: maxHeight,
           isOwnMessage: isOwnMessage,
           messageOptions: messageOptions,
           message: message,
           messageTextBuilder: (ChatMessage m, ChatMessage? p, ChatMessage? n) {
             return Row(
+              mainAxisSize: MainAxisSize.min,
               children: <Widget>[
                 Padding(
-                  padding: const EdgeInsets.only(right: 8.0),
+                  padding: const EdgeInsets.only(right: 18.0),
                   child: !media.isUploading
                       ? Icon(
                           Icons.description,
@@ -96,49 +149,97 @@ class MediaContainer extends StatelessWidget {
   Widget build(BuildContext context) {
     if (message.medias != null && message.medias!.isNotEmpty) {
       final List<ChatMedia> media = message.medias!;
-      return Wrap(
-        alignment: isOwnMessage ? WrapAlignment.end : WrapAlignment.start,
-        children: media.map(
-          (ChatMedia m) {
-            final double gallerySize =
-                (MediaQuery.of(context).size.width * 0.7) / 2 - 5;
-            final bool isImage = m.type == MediaType.image;
-            return Container(
-              color: Colors.transparent,
-              margin: const EdgeInsets.only(top: 5, right: 5),
-              width: media.length > 1 && isImage ? gallerySize : null,
-              height: media.length > 1 && isImage ? gallerySize : null,
-              constraints: BoxConstraints(
-                maxHeight: MediaQuery.of(context).size.height * 0.5,
-                maxWidth: MediaQuery.of(context).size.width * 0.7,
-              ),
-              child: GestureDetector(
-                onTap: messageOptions.onTapMedia != null
-                    ? () => messageOptions.onTapMedia!(m)
-                    : null,
-                child: ClipRRect(
-                  borderRadius: BorderRadius.circular(8.0),
-                  child: ColorFiltered(
-                    colorFilter: ColorFilter.mode(
-                      m.isUploading
-                          ? Colors.white54
-                          : Colors.white.withOpacity(
-                              0.1,
-                            ), // Because transparent is causing an issue on flutter web
-                      BlendMode.srcATop,
+      //Size size = MediaQuery.sizeOf(context);
+      double maxWidthPerc = maxWidth * 0.8;
+      double maxHeightPerc = maxHeight * 0.8;
+
+      return Container(
+        decoration: defaultMessageDecoration(
+          color: isOwnMessage
+              ? messageOptions.currentUserContainerColor(context)
+              : messageOptions.containerColor,
+          borderTopLeft:
+          isPreviousSameAuthor && !isOwnMessage && !isAfterDateSeparator
+              ? 0.0
+              : messageOptions.borderRadius,
+          borderTopRight:
+          isPreviousSameAuthor && isOwnMessage && !isAfterDateSeparator
+              ? 0.0
+              : messageOptions.borderRadius,
+          borderBottomLeft:
+          !isOwnMessage && !isBeforeDateSeparator && isNextSameAuthor
+              ? 0.0
+              : messageOptions.borderRadius,
+          borderBottomRight:
+          isOwnMessage && !isBeforeDateSeparator && isNextSameAuthor
+              ? 0.0
+              : messageOptions.borderRadius,
+        ),
+        padding: messageOptions.messagePadding,
+        child: Column(
+          crossAxisAlignment: isOwnMessage ? CrossAxisAlignment.end : CrossAxisAlignment.start,
+          children: <Widget>[
+            Wrap(
+              alignment: isOwnMessage ? WrapAlignment.end : WrapAlignment.start,
+              children: media.map(
+                (ChatMedia m) {
+                  final double gallerySize = ((maxWidth * 0.6) / 2 - 5).clampMax(maxWidthPerc);
+                  final bool isImage = m.type == MediaType.image;
+                  return Container(
+                    color: Colors.transparent,
+                    margin: const EdgeInsets.only(top: 5, right: 5),
+                    width: media.length > 1 && isImage ? gallerySize : null,
+                    height: media.length > 1 && isImage ? gallerySize : null,
+                    constraints: BoxConstraints(
+                      maxHeight: (maxHeight * 0.4).clampMax(maxHeightPerc),
+                      maxWidth: (maxWidth * 0.6).clampMax(maxWidthPerc),
                     ),
-                    child: _getMedia(
-                      context,
-                      m,
-                      media.length > 1 ? gallerySize : null,
-                      media.length > 1 ? gallerySize : null,
+                    child: ZoomTapContent(
+                      onTap: messageOptions.onTapMedia != null
+                          ? () => messageOptions.onTapMedia!(m)
+                          : null,
+                      child: ClipRRect(
+                        borderRadius: BorderRadius.circular(8.0),
+                        child: ColorFiltered(
+                          colorFilter: ColorFilter.mode(
+                            m.isUploading
+                                ? Colors.white54
+                                : Colors.white.withOpacity(
+                                    0.1,
+                                  ), // Because transparent is causing an issue on flutter web
+                            BlendMode.srcATop,
+                          ),
+                          child: _getMedia(
+                            context,
+                            m,
+                            media.length > 1 ? gallerySize : null,
+                            media.length > 1 ? gallerySize : null,
+                          ),
+                        ),
+                      ),
                     ),
+                  );
+                },
+              ).toList(),
+            ),
+            if (messageOptions.showTime)
+              messageOptions.messageTimeBuilder != null
+                  ? messageOptions.messageTimeBuilder!(message, isOwnMessage)
+                  : Padding(
+                padding: messageOptions.timePadding,
+                child: Text(
+                  (messageOptions.timeFormat ?? intl.DateFormat('HH:mm'))
+                      .format(message.createdAt),
+                  style: TextStyle(
+                    color: isOwnMessage
+                        ? messageOptions.currentUserTimeTextColor(context)
+                        : messageOptions.timeTextColor(),
+                    fontSize: messageOptions.timeFontSize,
                   ),
                 ),
               ),
-            );
-          },
-        ).toList(),
+          ],
+        ),
       );
     }
     return const SizedBox();